[FIX] sale_timesheet: multiple lines create new project
- Create 2 products A & B set as:
Service Invoicing Policy: Milestones (manually set quantities on order)
Service Tracking: Create a new project but no task
- Create a SO with A & B, validate
- Set a delivered quantity for both products
- Go to the project overview

Only the delivered quantity of a single SO line is retrieved in the 'To
invoice'.

It occurs because only a single SO line is linked to the project, while
both are expected to be linked to it.

opw-2003097

closes odoo/odoo#33815

Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>