[FIX] survey: fix assets forward-port
For saas-12.2 a file was added in survey assets with
https://github.com/odoo/odoo/commit/b3eeba8a96989f6af817880d3d8d6322286a069b
(even though the file itself was added in the repo with
https://github.com/odoo/odoo/commit/6588d914e6a4debcfc0c09a97e4781f744d8b0cc)

As a saas-12.1 fix, this was made:
https://github.com/odoo/odoo/commit/a540ebcd97da4dea9d812a7b00e0826202f8313e
That fix split the survey assets in two parts. After forward-port, the
file added in saas-12.2 ended up in the wrong part.

Bug found while working on task-1961045

closes odoo/odoo#33821

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>