[FIX] account: fix to_check warning being thrown when opening reconciliation widget
In the reconciliation widget, we added a to_check functionality. However in some cases, the propositions don't have a default value for
to_check, which result in a warning being thrown when doing a comparison to ensure that all lines have the same to_check value.
This commit add a default value to to_check for all propositions (when loading the data and when performing a quick create).

closes odoo/odoo#33824

Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>