[IMP] website_forum: display forum errors with notif
Previously the Forum errors were handled with a popover, now the
errors will be consistent with the crash_manager and display a toast
notification via the show_warning method.

Also the KarmaError were previously not displaying the karma needed to
perform the action. Now it will display the amount in the error message.

The CrashManager show_warning method will now take an options dictionary
as a second argument to override it's default behaviour.

Part of https://github.com/odoo/odoo/pull/32132
task-1894820